pathfinding: fix bug where node would not be found.
The search radius was set to k+snapFactorEdge to find nearby edges (bvh). This is wrong! The for loop might walk further away than k units (for example: distance(0,0 <-> 4,3) = 5, even tho k = 4!).

Because the for loop walked further away than anticipated, potential edges in range of snapFactorEdge could not be found. This can even lead to the case that a waypoint lands on a node, even if it was initially to for away.